Add new life form: WallFollowers
There is, now, a new critter that keeps it left hand to the wall and
follows along.

To implement this, the author of the book came up with `dirPlus`
function which calculates relative directions. So `dirPlus('n', 1)`
means one 45-degree turn clockwise from north, giving 'ne'.

Have a look at the example *plan* world and *~* creature.

TODO: implement A More Lifelike Simulation
https://eloquentjavascript.net/2nd_edition/07_elife.html#h_yTrS5ee3KB